Trait dioxus_lib::events::InteractionLocation

pub trait InteractionLocation {
    // Required methods
    fn client_coordinates(&self) -> Point2D<f64, ClientSpace>;
    fn screen_coordinates(&self) -> Point2D<f64, ScreenSpace>;
    fn page_coordinates(&self) -> Point2D<f64, PageSpace>;
}

A interaction that contains data about the location of the event.

Required Methods

fn client_coordinates(&self) -> Point2D<f64, ClientSpace>

Gets the coordinates of the event relative to the browser viewport.

fn screen_coordinates(&self) -> Point2D<f64, ScreenSpace>

Gets the coordinates of the event relative to the screen.

fn page_coordinates(&self) -> Point2D<f64, PageSpace>

Gets the coordinates of the event relative to the page.