Module dioxus_lib::prelude

Re-exports

pub use dioxus_html as dioxus_elements;

pub use dioxus_core;

Modules

keyboard_types

Contains types to define keyboard related events.

Macros

format_args_f

render

The rsx! macro makes it easy for developers to write jsx-style markup in their components.

rsx

The rsx! macro makes it easy for developers to write jsx-style markup in their components.

to_owned

A helper macro for using hooks and properties in async environments.

Structs

AnimationData

Attribute

An attribute on a DOM node, such as id="my-thing" or href="https://example.com"

ClipboardData

CompositionData

CopyValue

CopyValue is a wrapper around a value to make the value mutable and Copy.

Coroutine

DragData

The DragEvent interface is a DOM event that represents a drag and drop interaction. The user initiates a drag by placing a pointer device (such as a mouse) on the touch surface and then dragging the pointer to a new location (such as another DOM element). Applications are free to interpret a drag and drop interaction in an application-specific way.

ErrorBoundary

A boundary that will capture any errors from child components

Event

A wrapper around some generic data that handles the event’s state

EventHandler

The callback type generated by the rsx! macro when an on field is specified for components.

FocusData

FormData

GlobalMemo

A signal that can be accessed from anywhere in the application and created in a static

GlobalSignal

A signal that can be accessed from anywhere in the application and created in a static

ImageData

KeyboardData

MappedSignal

A read only signal that has been mapped to a new type.

MediaData

Modifiers

Pressed modifier keys.

MountedData

An Element that has been rendered and allows reading and modifying information about it.

MouseData

A synthetic event that wraps a web-style MouseEvent Data associated with a mouse event

Owner

Owner: Handles dropping generational boxes. The owner acts like a runtime lifetime guard. Any states that you create with an owner will be dropped when that owner is dropped.

PlatformEventData

A platform specific event.

PointerData

ReactiveContext

A context for signal reads and writes to be directed to

ReadOnlySignal

A signal that can only be read from.

ReadableValueIterator

An iterator over the values of a Readable<Vec<T>>.

Resource

Runtime

A global runtime that is shared across all scopes that provides the async runtime and context API

RuntimeGuard

A guard for a new runtime. This must be used to override the current runtime when importing components from a dynamic library that has it’s own runtime.

ScopeId

A component’s unique identifier.

ScopeState

A component’s rendered state.

ScrollData

SelectionData

SerializedAnimationData

A serialized version of AnimationData

SerializedClipboardData

A serialized version of ClipboardData

SerializedCompositionData

A serialized version of CompositionData

SerializedDragData

A serialized version of DragData

SerializedFocusData

A serialized version of FocusData

SerializedFormData

A serialized form data object

SerializedImageData

A serialized version of ImageData

SerializedKeyboardData

A serialized version of KeyboardData

SerializedMediaData

A serialized version of MediaData

SerializedMouseData

A serialized version of MouseData

SerializedPointerData

A serialized version of PointerData

SerializedScrollData

A serialized version of ScrollData

SerializedSelectionData

A serialized version of SelectionData

SerializedToggleData

A serialized version of ToggleData

SerializedTouchData

A serialized version of TouchData

SerializedTransitionData

A serialized version of TransitionData

SerializedWheelData

A serialized version of WheelData

Signal

Creates a new Signal. Signals are a Copy state management solution with automatic dependency tracking.

SignalData

The data stored for tracking in a signal.

SyncStorage

A thread safe storage. This is slower than the unsync storage, but allows you to share the value between threads.

Task

A task’s unique identifier.

Template

A static layout of a UI tree that describes a set of dynamic and static nodes.

ToggleData

TouchData

TouchPoint

TransitionData

UnboundedReceiver

The receiving end of an unbounded mpsc channel.

UnboundedSender

The transmission end of an unbounded mpsc channel.

UnsyncStorage

A unsync storage. This is the default storage type.

UseCallback

This callback is not generic over a return type so you can hold a bunch of callbacks at once

UseEval

A wrapper around the target platform’s evaluator.

UseFuture

VNode

A reference to a template along with any context needed to hydrate it

VNodeInner

A reference to a template along with any context needed to hydrate it

VirtualDom

A virtual node system that progresses user events and diffs UI trees.

WheelData

WritableValueIterator

An iterator over the values of a Writable<Vec<T>>.

Write

A mutable reference to a signal’s value.

Enums

Code

Code is the physical position of a key.

EvalError

Represents an error when evaluating JavaScript

FormValue

A form value that may either be a list of values or a single value

Key

Key represents the meaning of a keypress.

KeyCode

Location

The location attribute contains an indication of the logical location of the key on the device.

MountedError

The error type for the MountedData

ScrollBehavior

The way that scrolling should be performed

TemplateAttribute

An attribute of the TemplateNode, created at compile time

TemplateNode

A statically known node in a layout.

UseFutureState

A signal that represents the state of a future

UseResourceState

A signal that represents the state of a future

Traits

AExtension

AbbrExtension

AddressExtension

AnimateExtension

AnimateMotionExtension

AnimateTransformExtension

AnnotationExtension

AnnotationXmlExtension

AnyStorage

A trait for any storage backing type.

AreaExtension

ArticleExtension

AsideExtension

AudioExtension

BExtension

BaseExtension

BdiExtension

BdoExtension

BlockquoteExtension

BodyExtension

BrExtension

ButtonExtension

CanvasExtension

CaptionExtension

CircleExtension

CiteExtension

ClipPathExtension

CodeExtension

ColExtension

ColgroupExtension

ComponentFunction

Any component that implements the ComponentFn trait can be used as a component.

DataExtension

DatalistExtension

DdExtension

DefsExtension

DelExtension

Dep

A dependency is a trait that can be used to determine if a effect or selector should be re-run.

Dependency

A dependency is a trait that can be used to determine if a effect or selector should be re-run.

DescExtension

DetailsExtension

DfnExtension

DialogExtension

DiscardExtension

DivExtension

DlExtension

DtExtension

EllipseExtension

EmExtension

EmbedExtension

EvalProvider

A struct that implements EvalProvider is sent through ScopeState’s provide_context function so that [use_eval] can provide a platform agnostic interface for evaluating JavaScript code.

Evaluator

The platform’s evaluator.

FeBlendExtension

FeColorMatrixExtension

FeComponentTransferExtension

FeCompositeExtension

FeConvolveMatrixExtension

FeDiffuseLightingExtension

FeDisplacementMapExtension

FeDistantLightExtension

FeDropShadowExtension

FeFloodExtension

FeFuncAExtension

FeFuncBExtension

FeFuncGExtension

FeFuncRExtension

FeGaussianBlurExtension

FeImageExtension

FeMergeExtension

FeMergeNodeExtension

FeMorphologyExtension

FeOffsetExtension

FePointLightExtension

FeSpecularLightingExtension

FeSpotLightExtension

FeTileExtension

FeTurbulenceExtension

FieldsetExtension

FigcaptionExtension

FigureExtension

FilterExtension

FooterExtension

ForeignObjectExtension

FormExtension

GExtension

GlobalAttributes

GlobalAttributesExtension

H1Extension

H2Extension

H3Extension

H4Extension

H5Extension

H6Extension

HasAnimationData

A trait for any object that has the data for an animation event

HasAttributes

A trait for anything that has a dynamic list of attributes

HasClipboardData

HasCompositionData

A trait for any object that has the data for a composition event

HasDragData

A trait for any object that has the data for a drag event

HasFocusData

HasFormData

An object that has all the data for a form event

HasImageData

A trait for any object that has the data for an image event

HasKeyboardData

HasMediaData

HasMouseData

A trait for any object that has the data for a mouse event

HasPointerData

A trait for any object that has the data for a pointer event

HasScrollData

HasSelectionData

HasToggleData

HasTouchData

HasTouchPointData

A trait for touch point data

HasTransitionData

HasWheelData

HatchExtension

HatchpathExtension

HeadExtension

HeaderExtension

HgroupExtension

HrExtension

HtmlEventConverter

A converter between a platform specific event and a general event. All code in a renderer that has a large binary size should be placed in this trait. Each of these functions should be snipped in high levels of optimization.

IExtension

IframeExtension

ImgExtension

InputExtension

InsExtension

InteractionElementOffset

A interaction that contains data about the location of the event.

InteractionLocation

A interaction that contains data about the location of the event.

IntoAttributeValue

A value that can be converted into an attribute value

IntoDynNode

A trait that allows various items to be converted into a dynamic node for the rsx macro

KbdExtension

LabelExtension

LegendExtension

LiExtension

LineExtension

LinearGradientExtension

LinkExtension

MainExtension

MapExtension

MarkExtension

MarkerExtension

MaskExtension

MathExtension

MenuExtension

MerrorExtension

MetaExtension

MetadataExtension

MeterExtension

MfracExtension

MiExtension

MmultiscriptsExtension

MnExtension

MoExtension

ModifiersInteraction

A interaction that contains data about the current state of the keyboard modifiers.

MoverExtension

MpaddedExtension

MpathExtension

MphantomExtension

MprescriptsExtension

MrootExtension

MrowExtension

MsExtension

MspaceExtension

MsqrtExtension

MstyleExtension

MsubExtension

MsubsupExtension

MsupExtension

MtableExtension

MtdExtension

MtextExtension

MtrExtension

MunderExtension

MunderoverExtension

NavExtension

NoscriptExtension

ObjectExtension

OlExtension

OptgroupExtension

OptionExtension

OutputExtension

PExtension

ParamExtension

PathExtension

PatternExtension

PictureExtension

PointerInteraction

A interaction that contains data about the pointer button(s) that triggered the event.

PolygonExtension

PolylineExtension

PreExtension

ProgressExtension

Properties

Every “Props” used for a component must implement the Properties trait. This trait gives some hints to Dioxus on how to memoize the props and some additional optimizations that can be made. We strongly encourage using the derive macro to implement the Properties trait automatically as guarantee that your memoization strategy is safe.

QExtension

RadialGradientExtension

Readable

A trait for states that can be read from like crate::Signal, crate::GlobalSignal, or crate::ReadOnlySignal. You may choose to accept this trait as a parameter instead of the concrete type to allow for more flexibility in your API. For example, instead of creating two functions, one that accepts a crate::Signal and one that accepts a crate::GlobalSignal, you can create one function that accepts a Readable type.

ReadableOptionExt

An extension trait for Readable<Option> that provides some convenience methods.

ReadableVecExt

An extension trait for Readable<Vec> that provides some convenience methods.

RectExtension

RenderedElementBacking

An Element that has been rendered and allows reading and modifying information about it.

RpExtension

RtExtension

RubyExtension

SExtension

SampExtension

ScriptExtension

SectionExtension

SelectExtension

SemanticsExtension

SetExtension

SlotExtension

SmallExtension

SourceExtension

SpanExtension

StopExtension

Storage

A trait for a storage backing type. (RefCell, RwLock, etc.)

StrongExtension

StyleExtension

SubExtension

SummaryExtension

SupExtension

SuperFrom

A enhanced version of the From trait that allows with more flexibility.

SuperInto

A enhanced version of the Into trait that allows with more flexibility.

SvgAttributes

SvgAttributesExtension

SvgExtension

SwitchExtension

SymbolExtension

TableExtension

TbodyExtension

TdExtension

TemplateExtension

TextExtension

TextPathExtension

TextareaExtension

TfootExtension

ThExtension

TheadExtension

Throw

A trait to allow results to be thrown upwards to the nearest Error Boundary

TimeExtension

TitleExtension

TrExtension

TrackExtension

TspanExtension

UExtension

UlExtension

UseExtension

VarExtension

VideoExtension

ViewExtension

WbrExtension

Writable

A trait for states that can be read from like crate::Signal, or crate::GlobalSignal. You may choose to accept this trait as a parameter instead of the concrete type to allow for more flexibility in your API. For example, instead of creating two functions, one that accepts a crate::Signal and one that accepts a crate::GlobalSignal, you can create one function that accepts a Writable type.

WritableOptionExt

An extension trait for Writable<Option> that provides some convenience methods.

WritableVecExt

An extension trait for Writable<Vec> that provides some convenience methods.

Functions

ErrorBoundary

Create a new error boundary component.

Fragment

Create inline fragments using Component syntax.

consume_context

Consume context from the current scope

consume_context_from_scope

Consume context from the current scope

current_scope_id

Get the current scope id

eval

eval_provider

Get a struct that can execute any JavaScript.

event_bubbles

fc_to_builder

This utility function launches the builder method so rsx! and html! macros can use the typed-builder pattern to initialize a component’s props.

flush_sync

Wait for the virtualdom to finish its sync work before proceeding

generation

Get the current render since the inception of this component

has_context

Check if the current scope has a context

needs_update

Mark the current scope as dirty, causing it to re-render

needs_update_any

Mark the current scope as dirty, causing it to re-render

onabort

abort

onanimationend

onanimationend

onanimationiteration

onanimationiteration

onanimationstart

onanimationstart

onblur

onblur

oncanplay

canplay

oncanplaythrough

canplaythrough

onchange

onchange

onclick

Execute a callback when a button is clicked.

oncompositionend

oncompositionend

oncompositionstart

oncompositionstart

oncompositionupdate

oncompositionupdate

oncontextmenu

oncontextmenu

oncopy

oncopy

oncut

oncut

ondblclickDeprecated

ondoubleclick

ondrag

ondrag

ondragend

ondragend

ondragenter

ondragenter

ondragexit

ondragexit

ondragleave

ondragleave

ondragover

ondragover

ondragstart

ondragstart

ondrop

ondrop

ondurationchange

durationchange

onemptied

emptied

onencrypted

encrypted

onended

ended

onerror

onerror

onfocus

onfocus

onfocusin

onfocusout

ongotpointercapture

gotpointercapture

oninput

oninput handler

oninvalid

oninvalid

onkeydown

onkeydown

onkeypress

onkeypress

onkeyup

onkeyup

onload

onload

onloadeddata

loadeddata

onloadedmetadata

loadedmetadata

onloadstart

loadstart

onlostpointercapture

lostpointercapture

onmounted

mounted

onmousedown

onmousedown

onmouseenter

onmouseenter

onmouseleave

onmouseleave

onmousemove

onmousemove

onmouseout

onmouseout

onmouseover

onmouseover

onmouseup

onmouseup

onpaste

onpaste

onpause

pause

onplay

play

onplaying

playing

onpointercancel

pointercancel

onpointerdown

pointerdown

onpointerenter

pointerenter

onpointerleave

pointerleave

onpointermove

pointermove

onpointerout

pointerout

onpointerover

pointerover

onpointerup

pointerup

onprogress

progress

onratechange

ratechange

onreset

onreset

onscroll

onscroll

onseeked

seeked

onseeking

seeking

onselect

select

onselectionchange

selectionchange

onselectstart

selectstart

onstalled

stalled

onsubmit

onsubmit

onsuspend

suspend

ontimeupdate

timeupdate

ontoggle

ontoggle

ontouchcancel

touchcancel

ontouchend

touchend

ontouchmove

touchmove

ontouchstart

touchstart

ontransitionend

transitionend

onvolumechange

volumechange

onwaiting

waiting

onwheel

Called when the mouse wheel is rotated over an element.

parent_scope

Get the parent of the current scope if it exists

provide_context

Provide context to the current scope

provide_root_context

Provide a context to the root scope

remove_future

Informs the scheduler that this task is no longer needed and should be removed.

schedule_update

Schedule an update for the current component

schedule_update_any

Schedule an update for any component given its ScopeId.

set_event_converter

spawn

Spawns the future but does not return the [TaskId]

spawn_forever

Spawn a future that Dioxus won’t clean up when this component is unmounted

suspend

Suspends the current component

try_consume_context

Consume context from the current scope

try_use_context

Consume some context in the tree, providing a sharable handle to the value

use_after_render

Push this function to be run after the next render

use_before_render

A hook that allows you to insert a “before render” function.

use_callback

A callback that’s always current

use_context

Consume some context in the tree, providing a sharable handle to the value

use_context_provider

Provide some context via the tree and return a reference to it

use_coroutine

Maintain a handle over a future that can be paused, resumed, and canceled.

use_coroutine_handle

Get a handle to a coroutine higher in the tree

use_drop

Creates a callback that will be run before the component is removed. This can be used to clean up side effects from the component (created with use_effect).

use_effect

Create a new effect. The effect will be run immediately and whenever any signal it reads changes. The signal will be owned by the current component and will be dropped when the component is dropped.

use_error_boundary

Provide an error boundary to catch errors from child components

use_future

A hook that allows you to spawn a future

use_hook

Store a value between renders. The foundational hook for all other hooks.

use_hook_did_run

A hook that uses before/after lifecycle hooks to determine if the hook was run

use_hook_with_cleanup

Use a hook with a cleanup function

use_maybe_sync_memo

Creates a new Selector that may be sync. The selector will be run immediately and whenever any signal it reads changes.

use_maybe_sync_selector_with_dependencies

Creates a new Selector that may be sync with some local dependencies. The selector will be run immediately and whenever any signal it reads or any dependencies it tracks changes

use_memo

Creates a new unsync Selector. The selector will be run immediately and whenever any signal it reads changes.

use_memo_with_dependencies

Creates a new unsync Selector with some local dependencies. The selector will be run immediately and whenever any signal it reads or any dependencies it tracks changes

use_on_unmountDeprecated

use_resource

A memo that resolve to a value asynchronously.

use_root_context

use_signal

Creates a new Signal. Signals are a Copy state management solution with automatic dependency tracking.

use_signal_sync

Creates a new `Send + Sync`` Signal. Signals are a Copy state management solution with automatic dependency tracking.

with_owner

Run a closure with the given owner.

Type Aliases

AnimationEvent

ClipboardEvent

Component

A Component is a function that takes a [Scope] and returns an Element.

CompositionEvent

DragEvent

Element

An Element is a possibly-none VNode created by calling render on [Scope] or ScopeState.

FocusEvent

FormEvent

ImageEvent

KeyboardEvent

MediaEvent

MountedEvent

MountedResult

The MountedResult type for the MountedData

MouseEvent

PointerEvent

A synthetic event that wraps a web-style PointerEvent

ReadableRef

A reference to a value that can be read from.

ScrollEvent

SelectionEvent

SyncSignal

A signal that can safely shared between threads.

ToggleEvent

TouchEvent

TransitionEvent

WheelEvent

Attribute Macros

component

Streamlines component creation. This is the recommended way of creating components, though you might want lower-level control with more advanced uses.

inline_props

Derive props for a component within the component definition.

Derive Macros

Props