Struct dioxus_lib::prelude::GlobalSignal

pub struct GlobalSignal<T> { /* private fields */ }

A signal that can be accessed from anywhere in the application and created in a static

Implementations

impl<T> GlobalSignal<T>
where T: 'static,

pub const fn new(initializer: fn() -> T) -> GlobalSignal<T>

Create a new global signal with the given initializer.

pub fn signal(&self) -> Signal<T>

Get the signal that backs this global.

pub fn write(&self) -> Write<T>

Write this value

pub fn origin_scope(&self) -> ScopeId

Get the scope the signal was created in.

pub fn with_mut<O>(&self, f: impl FnOnce(&mut T) -> O) -> O

Run a closure with a mutable reference to the signal’s value. If the signal has been dropped, this will panic.

pub fn id(&self) -> GenerationalBoxId

Get the generational id of the signal.