Struct dioxus_lib::prelude::VirtualDom

pub struct VirtualDom { /* private fields */ }

A virtual node system that progresses user events and diffs UI trees.

Guide

Components are defined as simple functions that take [Scope] and return an Element.

#[derive(Props, PartialEq, Clone)]
struct AppProps {
    title: String
}

fn app(cx: AppProps) -> Element {
    rsx!(
        div {"hello, {cx.title}"}
    )
}

Components may be composed to make complex apps.

static ROUTES: &str = "";

#[component]
fn app(cx: AppProps) -> Element {
    rsx!(
        NavBar { routes: ROUTES }
        Title { "{cx.title}" }
        Footer {}
    )
}

#[component]
fn NavBar( routes: &'static str) -> Element {
    rsx! {
        div { "Routes: {routes}" }
    }
}

#[component]
fn Footer() -> Element {
    rsx! { div { "Footer" } }
}

#[component]
fn Title( children: Element) -> Element {
    rsx! {
        div { id: "title", {children} }
    }
}

To start an app, create a VirtualDom and call VirtualDom::rebuild to get the list of edits required to draw the UI.

let mut vdom = VirtualDom::new(app);
let edits = vdom.rebuild_to_vec();

To call listeners inside the VirtualDom, call VirtualDom::handle_event with the appropriate event data.

vdom.handle_event(event);

While no events are ready, call VirtualDom::wait_for_work to poll any futures inside the VirtualDom.

vdom.wait_for_work().await;

Once work is ready, call [VirtualDom::render_with_deadline] to compute the differences between the previous and current UI trees. This will return a [Mutations] object that contains Edits, Effects, and NodeRefs that need to be handled by the renderer.

let mutations = vdom.work_with_deadline(tokio::time::sleep(Duration::from_millis(100)));

for edit in mutations.edits {
    real_dom.apply(edit);
}

To not wait for suspense while diffing the VirtualDom, call VirtualDom::render_immediate or pass an immediately ready future to [VirtualDom::render_with_deadline].

Building an event loop around Dioxus:

Putting everything together, you can build an event loop around Dioxus by using the methods outlined above.

#[component]
fn app() -> Element {
    rsx! {
        div { "Hello World" }
    }
}

let dom = VirtualDom::new(app);

real_dom.apply(dom.rebuild());

loop {
    select! {
        _ = dom.wait_for_work() => {}
        evt = real_dom.wait_for_event() => dom.handle_event(evt),
    }

    real_dom.apply(dom.render_immediate());
}

Waiting for suspense

Because Dioxus supports suspense, you can use it for server-side rendering, static site generation, and other usecases where waiting on portions of the UI to finish rendering is important. To wait for suspense, use the [VirtualDom::render_with_deadline] method:

let dom = VirtualDom::new(app);

let deadline = tokio::time::sleep(Duration::from_millis(100));
let edits = dom.render_with_deadline(deadline).await;

Use with streaming

If not all rendering is done by the deadline, it might be worthwhile to stream the rest later. To do this, we suggest rendering with a deadline, and then looping between VirtualDom::wait_for_work and render_immediate until no suspended work is left.

let dom = VirtualDom::new(app);

let deadline = tokio::time::sleep(Duration::from_millis(20));
let edits = dom.render_with_deadline(deadline).await;

real_dom.apply(edits);

while dom.has_suspended_work() {
   dom.wait_for_work().await;
   real_dom.apply(dom.render_immediate());
}

Implementations

impl VirtualDom

pub fn new(app: fn() -> Option<VNode>) -> VirtualDom

Create a new VirtualDom with a component that does not have special props.

Description

Later, the props can be updated by calling “update” with a new set of props, causing a set of re-renders.

This is useful when a component tree can be driven by external state (IE SSR) but it would be too expensive to toss out the entire tree.

Example

fn Example() -> Element  {
    rsx!( div { "hello world" } ))
}

let dom = VirtualDom::new(Example);

Note: the VirtualDom is not progressed, you must either “run_with_deadline” or use “rebuild” to progress it.

pub fn new_with_props<P, M>( root: impl ComponentFunction<P, M>, root_props: P ) -> VirtualDom
where P: Clone + 'static, M: 'static,

Create a new VirtualDom with the given properties for the root component.

Description

Later, the props can be updated by calling “update” with a new set of props, causing a set of re-renders.

This is useful when a component tree can be driven by external state (IE SSR) but it would be too expensive to toss out the entire tree.

Example

#[derive(PartialEq, Props)]
struct SomeProps {
    name: &'static str
}

fn Example(cx: SomeProps) -> Element  {
    rsx!{ div { "hello {cx.name}" } }
}

let dom = VirtualDom::new(Example);

Note: the VirtualDom is not progressed on creation. You must either “run_with_deadline” or use “rebuild” to progress it.

let mut dom = VirtualDom::new_with_props(Example, SomeProps { name: "jane" });
let mutations = dom.rebuild();

pub fn prebuilt(app: fn() -> Option<VNode>) -> VirtualDom

Create a new virtualdom and build it immediately

pub fn get_scope(&self, id: ScopeId) -> Option<&ScopeState>

Get the state for any scope given its ID

This is useful for inserting or removing contexts from a scope, or rendering out its root node

pub fn base_scope(&self) -> &ScopeState

Get the single scope at the top of the VirtualDom tree that will always be around

This scope has a ScopeId of 0 and is the root of the tree

pub fn in_runtime<O>(&self, f: impl FnOnce() -> O) -> O

Run a closure inside the dioxus runtime

pub fn with_root_context<T>(self, context: T) -> VirtualDom
where T: Clone + 'static,

Build the virtualdom with a global context inserted into the base scope

This is useful for what is essentially dependency injection when building the app

pub fn insert_any_root_context(&mut self, context: Box<dyn Any>)

Build the virtualdom with a global context inserted into the base scope

This method is useful for when you want to provide a context in your app without knowing its type

pub fn mark_dirty(&mut self, id: ScopeId)

Manually mark a scope as requiring a re-render

Whenever the Runtime “works”, it will re-render this scope

pub fn handle_event( &mut self, name: &str, data: Rc<dyn Any>, element: ElementId, bubbles: bool )

Call a listener inside the VirtualDom with data from outside the VirtualDom. The ElementId passed in must be the id of an element with a listener, not a static node or a text node.

This method will identify the appropriate element. The data must match up with the listener declared. Note that this method does not give any indication as to the success of the listener call. If the listener is not found, nothing will happen.

It is up to the listeners themselves to mark nodes as dirty.

If you have multiple events, you can call this method multiple times before calling “render_with_deadline”

pub async fn wait_for_work(&mut self)

Wait for the scheduler to have any work.

This method polls the internal future queue, waiting for suspense nodes, tasks, or other work. This completes when any work is ready. If multiple scopes are marked dirty from a task or a suspense tree is finished, this method will exit.

This method is cancel-safe, so you’re fine to discard the future in a select block.

This lets us poll async tasks and suspended trees during idle periods without blocking the main thread.

Example

let dom = VirtualDom::new(app);

pub fn process_events(&mut self)

Process all events in the queue until there are no more left

pub fn replace_template(&mut self, template: Template)

Replace a template at runtime. This will re-render all components that use this template. This is the primitive that enables hot-reloading.

The caller must ensure that the template references the same dynamic attributes and nodes as the original template.

This will only replace the the parent template, not any nested templates.

pub fn rebuild_in_place(&mut self)

Rebuild the virtualdom without handling any of the mutations

This is useful for testing purposes and in cases where you render the output of the virtualdom without handling any of its mutations.

pub fn rebuild_to_vec(&mut self) -> Mutations

VirtualDom::rebuild to a vector of mutations for testing purposes

pub fn rebuild(&mut self, to: &mut impl WriteMutations)

Performs a full rebuild of the virtual dom, returning every edit required to generate the actual dom from scratch.

The mutations item expects the RealDom’s stack to be the root of the application.

Tasks will not be polled with this method, nor will any events be processed from the event queue. Instead, the root component will be ran once and then diffed. All updates will flow out as mutations.

All state stored in components will be completely wiped away.

Any templates previously registered will remain.

Example

static app: Component = |cx|  rsx!{ "hello world" });

let mut dom = VirtualDom::new();
let edits = dom.rebuild();

apply_edits(edits);

pub fn render_immediate(&mut self, to: &mut impl WriteMutations)

Render whatever the VirtualDom has ready as fast as possible without requiring an executor to progress suspended subtrees.

pub fn render_immediate_to_vec(&mut self) -> Mutations

Self::render_immediate to a vector of mutations for testing purposes

pub async fn wait_for_suspense(&mut self)

Render the virtual dom, waiting for all suspense to be finished

The mutations will be thrown out, so it’s best to use this method for things like SSR that have async content

We don’t call “flush_sync” here since there’s no sync work to be done. Futures will be progressed like usual, however any futures wating on flush_sync will remain pending

pub fn runtime(&self) -> Rc<Runtime>

Get the current runtime